requirements.txt
# ARIA Phase 1 — Dependencies
# Core async HTTP
aiohttp>=3.9.0

# Web scraping (GitHub Trending)
beautifulsoup4>=4.12.0

# RSS parsing (AWS Blog)
feedparser>=6.0.0

# ArXiv API XML parsing
xmltodict>=0.13.0

# Tavily search
tavily-python>=0.3.0

# Environment variables
python-dotenv>=1.0.0

# Date/time handling
python-dateutil>=2.8.0
